+++ bt-core/src/main/java/bt/data/DataDescriptorFactory.java
@@ -24,17 +24,22 @@ import bt.metainfo.Torrent;
  */
 public class DataDescriptorFactory implements IDataDescriptorFactory {
 
-    private ChunkVerifier verifier;
-    private int transferBlockSize;
+    private final DataReaderFactory dataReaderFactory;
+    private final ChunkVerifier verifier;
+    private final int transferBlockSize;
 
-    public DataDescriptorFactory(ChunkVerifier verifier,
-                                 int transferBlockSize) {
+    public DataDescriptorFactory(
+            DataReaderFactory dataReaderFactory,
+            ChunkVerifier verifier,
+            int transferBlockSize) {
+
+        this.dataReaderFactory = dataReaderFactory;
         this.verifier = verifier;
         this.transferBlockSize = transferBlockSize;
     }
 
     @Override
     public DataDescriptor createDescriptor(Torrent torrent, Storage storage) {
-        return new DefaultDataDescriptor(storage, torrent, verifier, transferBlockSize);
+        return new DefaultDataDescriptor(storage, torrent, verifier, dataReaderFactory, transferBlockSize);
     }
 }