+++ bt-core/src/main/java/bt/processor/torrent/InitializeTorrentProcessingStage.java
@@ -66,7 +66,7 @@ public class InitializeTorrentProcessingStage<C extends TorrentContext> extends 
 
         context.getRouter().registerMessagingAgent(GenericConsumer.consumer());
         context.getRouter().registerMessagingAgent(new BitfieldConsumer(bitfield, pieceStatistics, eventSink));
-        context.getRouter().registerMessagingAgent(new PieceConsumer(bitfield, dataWorker));
+        context.getRouter().registerMessagingAgent(new PieceConsumer(bitfield, dataWorker, eventSink));
         context.getRouter().registerMessagingAgent(new PeerRequestConsumer(dataWorker));
         context.getRouter().registerMessagingAgent(new RequestProducer(descriptor.getDataDescriptor()));
         context.getRouter().registerMessagingAgent(new MetadataProducer(() -> context.getTorrent().orElse(null), config));