+++ RELEASE-NOTES.md
@@ -10,6 +10,7 @@ For the latest information visit project web site: http://atomashpolskiy.github.
 
 * UPnP port mapping [#80](https://github.com/atomashpolskiy/bt/issues/80)
 * Generate OSGi manifests
+* Add bt.data.DataDescriptor.getReader() method for convenient stream-like access to torrent's data
 
 ### Bug Fixes/Improvements: